refactor: Mod disease detail util screen
Refactor disease detail view to new util screen content
Apply util disease detail screen on detail view and result disease view

Related to: #21
